Refactor HeaderLayout component and rename CommonSlider
Refactored the HeaderLayout component to move the menuData outside the component for better readability and efficiency. The CommonSlider component was renamed to EnhancedSwiper to better reflect its functionality. Some properties and methods were also changed in the Swiper component for better integration with data supplied. This allows more flexibility when using the Swiper component as it can now adapt to various data formats.