Refactor EnhancedSwiper for flexibility and added functionalities
The EnhancedSwiper component was considerably refactored to improve its flexibility & allow for additional features. Additional props were added to make the component more versatile, enabling configuration of individual slide components and their properties. To achieve this, the 'SlideData' type was added, enhancing the types and structure of data each slide can take.

In the spirit of providing better customization, the 'Slots' prop was introduced. This feature provides optional areas at the start/end of the 'wrapper' and 'container' where custom content can be added. These changes were done to make the EnhancedSwiper component more easily configurable and adaptable to wider use-cases.

Lastly, the previous static SwiperSlide was replaced with a dynamic MaterialSwiperSlide for a more performant and stylable swiper slide. The change also makes the component styles more consistent with the Material-UI styles/theme.

All these alterations were performed to enhance the usability and adaptability of the EnhancedSwiper component, while ensuring a more consistent project style guide and performance optimization.