Remove EnhancedSwiper from HomePage
EnhancedSwiper component was removed from the HomePage for performance optimization and to reduce unnecessary complexity. Homepage now only renders a single button, reducing loading times and providing a more streamlined user experience.

In the previous version, the page was using the EnhancedSwiper that included a static slide with set data. This data was redundant and not being used elsewhere in the application, resulting in unnecessary resource consumption. After examining the use-cases, it was evident the page didn't require a swiper for it's intended function, leading to the removal decision. Future components can be added when additional functionality for HomePage is required.