Add framer-motion & export MaterialSwiper and MaterialSwiperSlide
Added framer-motion to package.json and package-lock.json as it is necessary for animation capabilities in the application. Additionally, MaterialSwiper and MaterialSwiperSlide were exported from the enhanced-swiper.tsx file to allow for their usage in other parts of the application. Now these components can be utilized anywhere in the application enabling more efficient and reusable code.