Refactor m-link.tsx and implement eslint-config-prettier
Simplified the code in m-link.tsx file and exported as MLink for semantic clarity and easier usability. Replaced 'CSLink' with 'MLink' in all places where 'CSLink' was used to maintain consistency.

Integrated eslint-config-prettier into the .eslintrc.json and package.json to ensure prettier rules override ESLint rules that can potentially conflict with prettier. This helps maintain uniformity and consistency in the code formatting across the application.

Changed a 'Button' component to 'MLink' in the 'HomePage' component in page.tsx, ensuring navigation links are consistent throughout the application.