Update MLink component, add AnimatedButton and adjust Eslint config
Updated the underlying structure of 'MLink' component in m-link.tsx to use 'forwardRef' for better control and access to the DOM element. Added a displayName to help in debugging. Brought minor changes in Eslint configuration in .eslintrc.json for improved formatting and readability.

Moreover, we added a new component 'AnimatedButton' with animations to enhance UI/UX. Made necessary changes to button.theme.ts in theme styling to accommodate the new component.

The HomePage content is also updated to incorporate usage of new 'AnimatedButton', thereby enhancing the overall look and feel of the application.