Cleanup Page.tsx and fix link style in caseStudyCard.tsx
Removed unnecessary imports and components from Page.tsx to simplify the file structure. Logo and MLink were not required on the page and caused clutter. Also, modified the MLink style in caseStudyCard.tsx to improve consistency across the app. Instead of specifying the typography and color on each use, the styling is now handled globally. This change reduces redundancy and makes the codebase easier to manage.