Update Header and Navigation for dynamic routing and maintainability
In this commit, dynamic routing has been introduced to the Header component by passing route data as props, improving flexibility and promoting code reuse. Additionally, the Navigation component has been refactored to include separate 'DesktopNavigation', making the code cleaner and more maintainable. The update also involves removal of redundant paths and improvements on naming for the sake of clarity and simplicity. Lastly, a minor bug which was causing an incorrect display of background color in the tooltip of NavItem has been rectified.