Update player model with a new attribute - stamina
The Player model has been updated with a new attribute named 'stamina' to reflect the new game logic. The 'stamina' attribute will help track the amount of energy a player has left for performing actions in the game. This change was necessary to balance gameplay and add an extra layer of strategy and skill to the mix, as players will now have to manage their stamina consumption carefully.