Improve logo animation and header responsiveness
Removed unused 'useMediaQuery' from the 'navigation-action' component and added a conditional rendering to 'NavigationDrawer' to enhance the responsiveness of the navigation drawer. Updated the 'NavigationDrawer' component to use 'appbarHeight' for the top margin to better align it with the header. Created a responsive height constant 'appbarHeight' to adjust the header height based on the screen size. These changes significantly improve the layout on different devices and enhance the user experience overall.