Add hide-on-scroll functionality to AppBar
The AppBar component in 'index.tsx' has been wrapped in a new HideOnScroll function to hide it when the user scrolls down, and show it again when they scroll up. This improves use of screen real estate by only showing the AppBar when it's needed.
The HideOnScroll uses the Material-UI's 'useScrollTrigger' hook and 'Slide' transition for smooth, responsive reveal and hide behaviors. This includes import changes to accommodate those used functions and typing from 'react' for children of HideOnScroll.