Refactor drawer and backdrop styling
Enhanced the drawer and backdrop components' styles to improve user readability and aesthetic appeal. Background color and backdrop filter of drawer and dialog converted to a more appealing appearance. Removed unused styles in 'backdropVariants.ts' and 'nav-item.tsx'. The overall aim of these changes is to provide a better user experience and a more consistent and visually pleasing interface.