Add SectionProps to SectionWrapper component
Extended SectionWrapperProps interface to include SectionProps, which allows additional flexibility when using this component. It enables users to pass additional or override the default props of the Stack component being used internally. The main motivation behind this change was to provide the facility to control section-based styles or props at the usage level. Moreover, this change respects our effort to maintain generic components with adjustable prop values.