Revise naming and structure in various code files
The code refactoring was mainly focused on improving naming conventions and modularity, which would ultimately lead to better maintainability and readability of the code. Renamed several components from camelCase to kebab-case, and streamlined the handling of section wrappers. Introduced 'use client' in various files which was missing before, leading to better efficiency. Created a new file 'our-stories.tsx' to handle a specific feature separately. Made appropriate changes in 'page.tsx' to reflect the renamed components. Also, the 'ReactiveAppBar' logic was moved from 'index.tsx' to a new separate file 'react