Remove unused context and refactor EnhancedSwiper
Deleted 'context.ts' as it is not used anywhere in the application, clearing unnecessary clutter. Fixed a typo in 'key-focus.tsx' file for better readability. In 'enhanced-swiper.tsx', updated the types in the EnhancedSwiperProps interface for more flexibility. Unused code in 'key-focus.tsx' and 'smallTitleCard.tsx' was also cleaned up. These changes simplify the codebase and improve maintainability.