Remove mobile menu component and static data file
This commit removes the mobile menu component and the static data file that was providing the structure for the mobile menu. This was done because the mobile menu implementation was redundant, and removing it helps to simplify the codebase.

The diff also shows minor rearrangement in the packages included in the package.json file. This doesn't introduce any functional changes, it simply reorders the appearance of 'recoil' and 'swiper' in the file. No new package is introduced nor are any removed.