Refactor motionElements to motion-elements file
I have renamed motionElements.ts file into motion-elements.ts for maintaining the codebase in an easier-to-read and easier-to-maintain format. Along with that, I have shifted the import statements from the commonComponents/motionElements to commonComponents/motion-elements in multiple components. In addition, we now have added Grid2 component to our motion. This functional change enhances motion effects for grid components.